refactor: split raes/module_registry into an API-stable subdomain package
