+++ tools/policy/oversized_allowlist.yaml
@@ -17,7 +17,6 @@ files:
   - implementations/python/packages/raes_runtime/control_plane_api.py
   - implementations/python/packages/raes_runtime/workflow_result_contract_checks.py
   - implementations/python/packages/raes/composition.py
-  - implementations/python/packages/raes/module_registry.py
   - implementations/python/packages/raes/orchestration.py
   - implementations/python/packages/raes/participant_behavior.py
   - implementations/python/packages/raes/runtime_mail_service.py
